KeyFile: compare the OnlyShowIn/NotShowIn keys with XDG_CURRENT_DESKTOP
update the show property to use the XDG_CURRENT_DESKTOP environment
instead of DESKTOP_SESSION, if none of the current desktop have a
match, we default to show if the OnlyShowIn key is not present.

this is the proper behaviour defined in the desktop entry spec, and
the autostart doesn't differ here.